Use PostgreSQL in production.
This is as described in
https://www.railstutorial.org/book#sec-heroku_commands_setup, which
recommends using SQLite in development and test but PostgreSQL in
production.  This is for Heroku's sake, since Heroku does not support
SQLite.

This is in conflict with a principle of simplicity: minimize differences
between development, test, and production.  This is called out in Heroku's
own documentation here
https://devcenter.heroku.com/articles/getting-started-with-rails6.

We will go with the Ruby on Rails Tutorial for now, mainly because this
lets me avoid configuring PostgreSQL locally and in test, but I don't plan
to continue with this long term.